requirement.txt
pytest==7.2.0
selenium==4.5.0
PyMySQL==1.0.2
openpyxl==3.0.10
pandas==1.5.1
allure-pytest==2.11.1
pytest-playwright==0.3.0
python-dotenv-yaml==0.17.1
python-dotenv==0.21.0
urllib3==1.26.13
webdriver-manager==3.8.5
requests==2.28.1
openpyxl==3.0.10